Move code out of LuceneIndexMaintenanceTest into a new class
